backend/inspect_adk.py
import google.adk.tools as tools_module

print("All exports from google.adk.tools:")
for item in dir(tools_module):
    if not item.startswith('_'):
        obj = getattr(tools_module, item)
        print(f"  - {item}: {type(obj)}")
